+++ index.js
@@ -17,6 +17,7 @@ const make = (name, options) => ([
 export const shared = () => [{ dir: join(__dirname, '_shared', 'openapi'), ext: '@' }]
 export const singleUser = options => make('single-user', options)
 export const teamApiTokens = options => make('team-api-tokens', options)
+export const teamTasks = options => make('team-tasks', options)
 export const teams = options => make('teams', options)
 export const userApiTokens = options => make('user-api-tokens', options)
 export const userManagement = options => make('user-management', options)
@@ -27,6 +28,7 @@ export const all = options => ([
 	...shared(),
 	...singleUser(options),
 	...teamApiTokens(options),
+	...teamTasks(options),
 	...teams(options),
 	...userApiTokens(options),
 	...userManagement(options),